pyproject.toml
[build-system]
requires = ["setuptools==65.5.0", "wheel", "tomli"]
build-backend = "setuptools.build_meta"

[project]
name = "ube_mbrl"
description = "Model-Based Uncertainty in Value Functions"
version = "0.1"
authors = [
  {name = "Carlos Luis", email = "carlosenrique.luisgoncalves@de.bosch.com"},
]
classifiers=[
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
]
requires-python = ">=3.8"
dependencies = [
    "ipython~=8.1.0",
    "ipykernel~=6.15.2",
    "ipympl~=0.8.8",
    "matplotlib~=3.5.1",
    "black~=22.3",
    "numpy~=1.22",
    "torch==1.11.0",
    "scipy~=1.8",
    "pybullet~=3.2.5",
    "imageio-ffmpeg~=0.4.7",
    "swig",
    "pyglet==1.5.27",
    "gym[classic_control, box2d]==0.21.0",
    "dm_control~=1.0.7",
    "rlberry~=0.3.0",
    "pyvirtualdisplay~=3.0",
    # mbrl-lib
    "hydra-core==1.0.3",
    "termcolor>=1.1.0",
    "imageio>=2.9.0",
    "tqdm>=4.48.2",
]
dynamic = ["readme"]

[tool.flake8]
max-line-length = 100

[project.urls]
homepage = "https://github.com/boschresearch/ube-mbrl"
repository = "https://github.com/boschresearch/ube-mbrl"

[tool.setuptools.packages.find]
include = ["ube_mbrl*"]

[tool.setuptools.dynamic]
readme = {file = "README.md"}
